localtoc: get \@currentlocallevel from aux file
This fixes a bug in \sidetoc. Previuosly, the \@currentlocallevel was
set in the heading/preinit hook of dispositions. However, if the
notecolumn was processed *after* starting a new section or chapter,
the locallevel was wrong, leading to an empty localtoc. The fix
consists in grabbing the \@currentlocallevel from the previous entry
in the aux file. This means that we need to have starred entries
in the aux file, and they have the appropriate toc level.